+++ CHANGELOG.md
@@ -4,7 +4,13 @@
 
 ### New Features
 
+ - add new scope "instruction" for matching mnemonics and operands #767 @williballenthin
+ - add new feature "operand[{0, 1, 2}].number" for matching instruction operand immediate values #767 @williballenthin
+ - add new feature "operand[{0, 1, 2}].offset" for matching instruction operand offsets #767 @williballenthin
+
 ### Breaking Changes
+
+  - instruction scope and operand feature are new and are not backwards compatible with older versions of capa
 
 ### New Rules (4)